pyproject.toml
[tool.poetry]
name = "task-manager"
version = "0.1.0"
description = ""
authors = ["pranjal <rktpranjal@gmail.com>"]
readme = "README.md"
packages = [{include = "task_manager"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.118.0"
jupyter = "^1.1.1"
influxdb = "^5.3.2"
sqlalchemy = "^2.0.43"
python-dotenv = "^1.1.1"
redis = "^6.4.0"
psycopg-pool = "^3.2.6"
psycopg = {extras = ["binary"], version = "^3.2.10"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
